feat: :zap: move validation handler to core
