+++ api/src/test-util/testConfig.ts
@@ -13,6 +13,7 @@ import { TestTeamProvider } from "./TestTeamProvider.js";
 import { testUserProvider } from "./sampleUser.js";
 import { testSystemProvider } from "./system.js";
 import { StaticValidationProvider } from "@gram/config/dist/providers/static/StaticValidationProvider.js";
+import { testValidationRules } from "./testValidationRules.js";
 
 function toComponentClass(o: any): ComponentClass {
   return {
@@ -163,7 +164,7 @@ export const testConfig: GramConfiguration = {
       systemProvider: testSystemProvider,
       suggestionSources: [],
       teamProvider: new TestTeamProvider(),
-      validationProviders: [new StaticValidationProvider(dal)],
+      validationSources: [testValidationRules],
     };
   },
 };
