+++ core/src/config/GramConfiguration.ts
@@ -14,7 +14,7 @@ import type { Migration } from "../data/Migration.js";
 import type { TeamProvider } from "../auth/TeamProvider.js";
 import type { ActionItemExporter } from "../action-items/ActionItemExporter.js";
 import type { SearchProvider } from "../search/SearchHandler.js";
-import { ValidationProvider } from "../validation/ValidationHandler.js";
+import { ValidationRule } from "../validation/models.js";
 import type { DynamicAttribute } from "../attributes/attributes.js";
 
 export interface Providers {
@@ -38,7 +38,7 @@ export interface Providers {
   teamProvider?: TeamProvider;
   actionItemExporters?: ActionItemExporter[];
   searchProviders?: SearchProvider[];
-  validationProviders?: ValidationProvider[];
+  validationSources?: ValidationRule[][];
 }
 
 export interface GramConfiguration {
